Add files via upload
The technical document emphasizes the importance of the Certificate Signing Request (CSR) and the Certificate Key in the context of digital certificates and secure communication. The CSR is crucial as it includes entity details and the public key, facilitating identification verification and providing certificate attributes like the Distinguished Name (DN). The Certificate Key, which consists of private and public key pairs, plays a vital role in establishing secure communication, certifying trust and authentication, and creating a certificate binding. The private key is kept secret and used for decryption, while the public key is openly shared for verification and encryption purposes. Both keys contribute to the secure issuance and handling of certificates and have been instrumental in achieving the desired outcomes and knowledge acquisition in the portfolio.